LPS-161876 Rename the project
