--- modules/util/language-security-scan/src/main/resources/META-INF/maven/com.liferay/com.liferay.language.security.scan/plugin-help.xml
@@ -1,45 +1,0 @@
-<?xml version="1.0"?>
-
-<!-- Generated by maven-plugin-tools 3.5 -->
-
-<plugin>
-	<name>com.liferay.language.security.scan</name>
-	<description/>
-	<groupId>com.liferay</groupId>
-	<artifactId>com.liferay.language.security.scan</artifactId>
-	<version>1.0.1</version>
-	<goalPrefix>language-security-scan</goalPrefix>
-	<mojos>
-		<mojo>
-			<goal>format</goal>
-			<description>Invoke Liferay Language Security Scan to validate language property files.</description>
-			<requiresDirectInvocation>false</requiresDirectInvocation>
-			<requiresProject>true</requiresProject>
-			<requiresReports>false</requiresReports>
-			<aggregator>false</aggregator>
-			<requiresOnline>false</requiresOnline>
-			<inheritedByDefault>true</inheritedByDefault>
-			<implementation>com.liferay.language.security.scan.maven.LanguageSecurityScanMojo</implementation>
-			<language>java</language>
-			<instantiationStrategy>per-lookup</instantiationStrategy>
-			<executionStrategy>once-per-session</executionStrategy>
-			<threadSafe>false</threadSafe>
-			<parameters>
-				<parameter>
-					<name>baseDirName</name>
-					<type>java.lang.String</type>
-					<required>false</required>
-					<editable>true</editable>
-					<description/>
-				</parameter>
-				<parameter>
-					<name>scanAll</name>
-					<type>boolean</type>
-					<required>false</required>
-					<editable>true</editable>
-					<description/>
-				</parameter>
-			</parameters>
-		</mojo>
-	</mojos>
-</plugin>
