--- modules/util/language-security-scan/src/main/resources/META-INF/maven/plugin.xml
@@ -1,173 +1,0 @@
-<?xml version="1.0"?>
-
-<!-- Generated by maven-plugin-tools 3.5 -->
-
-<plugin>
-	<name>com.liferay.language.security.scan</name>
-	<description/>
-	<groupId>com.liferay</groupId>
-	<artifactId>com.liferay.language.security.scan</artifactId>
-	<version>1.0.1</version>
-	<goalPrefix>language-security-scan</goalPrefix>
-	<isolatedRealm>false</isolatedRealm>
-	<inheritedByDefault>true</inheritedByDefault>
-	<mojos>
-		<mojo>
-			<goal>format</goal>
-			<description>Invoke Liferay Language Security Scan to validate language property files.</description>
-			<requiresDirectInvocation>false</requiresDirectInvocation>
-			<requiresProject>true</requiresProject>
-			<requiresReports>false</requiresReports>
-			<aggregator>false</aggregator>
-			<requiresOnline>false</requiresOnline>
-			<inheritedByDefault>true</inheritedByDefault>
-			<implementation>com.liferay.language.security.scan.maven.LanguageSecurityScanMojo</implementation>
-			<language>java</language>
-			<instantiationStrategy>per-lookup</instantiationStrategy>
-			<executionStrategy>once-per-session</executionStrategy>
-			<threadSafe>false</threadSafe>
-			<parameters>
-				<parameter>
-					<name>baseDir</name>
-					<type>java.io.File</type>
-					<required>false</required>
-					<editable>false</editable>
-					<description/>
-				</parameter>
-				<parameter>
-					<name>baseDirName</name>
-					<type>java.lang.String</type>
-					<required>false</required>
-					<editable>true</editable>
-					<description/>
-				</parameter>
-				<parameter>
-					<name>scanAll</name>
-					<type>boolean</type>
-					<required>false</required>
-					<editable>true</editable>
-					<description/>
-				</parameter>
-			</parameters>
-			<configuration>
-				<baseDir implementation="java.io.File" default-value="${project.basedir}"/>
-			</configuration>
-		</mojo>
-	</mojos>
-	<dependencies>
-		<dependency>
-			<groupId>org.apache.ant</groupId>
-			<artifactId>ant</artifactId>
-			<type>jar</type>
-			<version>1.10.11</version>
-		</dependency>
-		<dependency>
-			<groupId>org.apache.ant</groupId>
-			<artifactId>ant-launcher</artifactId>
-			<type>jar</type>
-			<version>1.10.11</version>
-		</dependency>
-		<dependency>
-			<groupId>org.owasp.antisamy</groupId>
-			<artifactId>antisamy</artifactId>
-			<type>jar</type>
-			<version>1.7.1</version>
-		</dependency>
-		<dependency>
-			<groupId>net.sourceforge.htmlunit</groupId>
-			<artifactId>neko-htmlunit</artifactId>
-			<type>jar</type>
-			<version>2.64.0</version>
-		</dependency>
-		<dependency>
-			<groupId>org.apache.httpcomponents.client5</groupId>
-			<artifactId>httpclient5</artifactId>
-			<type>jar</type>
-			<version>5.1.3</version>
-		</dependency>
-		<dependency>
-			<groupId>org.apache.httpcomponents.core5</groupId>
-			<artifactId>httpcore5-h2</artifactId>
-			<type>jar</type>
-			<version>5.1.3</version>
-		</dependency>
-		<dependency>
-			<groupId>commons-codec</groupId>
-			<artifactId>commons-codec</artifactId>
-			<type>jar</type>
-			<version>1.15</version>
-		</dependency>
-		<dependency>
-			<groupId>org.apache.httpcomponents.core5</groupId>
-			<artifactId>httpcore5</artifactId>
-			<type>jar</type>
-			<version>5.1.4</version>
-		</dependency>
-		<dependency>
-			<groupId>org.apache.xmlgraphics</groupId>
-			<artifactId>batik-css</artifactId>
-			<type>jar</type>
-			<version>1.14</version>
-		</dependency>
-		<dependency>
-			<groupId>org.apache.xmlgraphics</groupId>
-			<artifactId>batik-shared-resources</artifactId>
-			<type>jar</type>
-			<version>1.14</version>
-		</dependency>
-		<dependency>
-			<groupId>org.apache.xmlgraphics</groupId>
-			<artifactId>batik-util</artifactId>
-			<type>jar</type>
-			<version>1.14</version>
-		</dependency>
-		<dependency>
-			<groupId>org.apache.xmlgraphics</groupId>
-			<artifactId>batik-constants</artifactId>
-			<type>jar</type>
-			<version>1.14</version>
-		</dependency>
-		<dependency>
-			<groupId>org.apache.xmlgraphics</groupId>
-			<artifactId>batik-i18n</artifactId>
-			<type>jar</type>
-			<version>1.14</version>
-		</dependency>
-		<dependency>
-			<groupId>org.apache.xmlgraphics</groupId>
-			<artifactId>xmlgraphics-commons</artifactId>
-			<type>jar</type>
-			<version>2.6</version>
-		</dependency>
-		<dependency>
-			<groupId>commons-io</groupId>
-			<artifactId>commons-io</artifactId>
-			<type>jar</type>
-			<version>2.11.0</version>
-		</dependency>
-		<dependency>
-			<groupId>org.slf4j</groupId>
-			<artifactId>slf4j-api</artifactId>
-			<type>jar</type>
-			<version>2.0.0</version>
-		</dependency>
-		<dependency>
-			<groupId>xerces</groupId>
-			<artifactId>xercesImpl</artifactId>
-			<type>jar</type>
-			<version>2.12.2</version>
-		</dependency>
-		<dependency>
-			<groupId>xml-apis</groupId>
-			<artifactId>xml-apis</artifactId>
-			<type>jar</type>
-			<version>1.4.01</version>
-		</dependency>
-		<dependency>
-			<groupId>xml-apis</groupId>
-			<artifactId>xml-apis-ext</artifactId>
-			<type>jar</type>
-			<version>1.3.04</version>
-		</dependency>
-	</dependencies>
-</plugin>
