--- modules/util/language-security-scan/src/main/resources/com/liferay/language/security/scan/ant/taskdefs.properties
@@ -1,1 +1,0 @@
-language-security-scan=com.liferay.language.security.scan.ant.LanguageSecurityScanTask
