+++ modules/util/language-security-scanner/build.gradle
@@ -1,0 +1,31 @@
+apply plugin: "com.liferay.maven.plugin.builder"
+
+buildPluginDescriptor {
+	ext {
+		autoClean = false
+	}
+
+	finalizedBy formatSource
+
+	goalPrefix = "language-security-scanner"
+}
+
+dependencies {
+	compile group: "org.apache.ant", name: "ant", version: "1.10.11"
+	compile group: "org.owasp.antisamy", name: "antisamy", version: "1.7.1"
+
+	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.impl", version: "2.0.0"
+	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "2.0.0"
+	compileOnly group: "org.apache.ant", name: "ant", transitive: false, version: "1.10.11"
+	compileOnly group: "org.apache.maven", name: "maven-plugin-api", transitive: false, version: "3.0.4"
+}
+
+liferay {
+	deployDir = "../../../tools/sdk/dependencies/com.liferay.language.security.scanner/lib"
+}
+
+updateFileVersions {
+	match(/\t\t\t\t<version>(\d.+)<\/version>/, "samples/pom.xml")
+	match(/com\.liferay\.language\.security\.scanner=(\d.+)/, "../../sdk/gradle-plugins/src/main/resources/com/liferay/gradle/plugins/dependencies/portal-tools.properties")
+	match(/\t\t<dependency name="com.liferay.language.security.scanner" org="com.liferay" rev="(\d.+)" \/>/, "../../../tools/sdk/dependencies/com.liferay.language.security.scanner/ivy.xml")
+}
