+++ modules/util/language-security-scanner/src/main/resources/antisamy-liferay.xml
