+++ modules/util/language-security-scanner/src/main/resources/com/liferay/language/security/scanner/ant/taskdefs.properties
@@ -1,0 +1,1 @@
+language-security-scanner=com.liferay.language.security.scanner.ant.LanguageSecurityScannerTask
