+++ tools/sdk/dependencies/com.liferay.language.security.scanner/ivy.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0"?>
+
+<ivy-module
+	version="2.0"
+	xmlns:m2="http://ant.apache.org/ivy/maven"
+	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:noNamespaceSchemaLocation="http://ant.apache.org/ivy/schemas/ivy.xsd"
+>
+	<info module="com.liferay.sdk.dependencies.com.liferay.language.security.scanner" organisation="com.liferay">
+		<extends extendType="configurations,description,info" location="${sdk.dir}/ivy.xml" module="com.liferay.sdk" organisation="com.liferay" revision="latest.integration" />
+	</info>
+
+	<dependencies defaultconf="default">
+		<dependency name="com.liferay.language.security.scanner" org="com.liferay" rev="1.0.0" />
+	</dependencies>
+</ivy-module>
